routes are pushed even when defaultGateway is true
The Office 365 on Windows hack requires a route to pushed in addition to the default gateway. The current configuration description suggests that this is not possible, but "routes" is still being used even if "defaultGateway" is set to true.

Therefore removing the statement that routes is only used when defaultGateway is false.